+++ Source/CesiumRuntime/Public/CesiumFeatureIdAttribute.h
@@ -28,8 +28,8 @@ enum class ECesiumFeatureIdAttributeStatus : uint8 {
 
 /**
  * @brief A blueprint-accessible wrapper for a feature ID attribute from a glTF
- * primitive. Provides access to per-vertex feature IDs which can be used with
- * the corresponding {@link FCesiumPropertyTable} to access per-vertex metadata.
+ * model. Provides access to feature IDs which can be used with the
+ * corresponding {@link FCesiumPropertyTable} to access metadata.
  */
 USTRUCT(BlueprintType)
 struct CESIUMRUNTIME_API FCesiumFeatureIdAttribute {
@@ -131,27 +131,27 @@ public:
       UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
-   * Get the number of vertices in the primitive containing the feature
-   * ID attribute. If the feature ID attribute is invalid, this returns 0.
+   * Get the number of feature IDs in this attribute. If the feature ID
+   * attribute is invalid, this returns 0.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Features|FeatureIDAttribute")
   static int64
-  GetVertexCount(UPARAM(ref)
-                     const FCesiumFeatureIdAttribute& FeatureIDAttribute);
+  GetFeatureIDCount(UPARAM(ref)
+                        const FCesiumFeatureIdAttribute& FeatureIDAttribute);
 
   /**
-   * Gets the feature ID associated with the given vertex. The feature ID can be
-   * used with a FCesiumFeatureTable to retrieve the per-vertex metadata. If
-   * the feature ID attribute is invalid, this returns -1.
+   * Gets the feature ID at the given index. A feature ID can be used with a
+   * FCesiumPropertyTable to retrieve the metadata for that ID. If the feature
+   * ID attribute is invalid, this returns -1.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Features|FeatureIDAttribute")
-  static int64 GetFeatureIDForVertex(
+  static int64 GetFeatureID(
       UPARAM(ref) const FCesiumFeatureIdAttribute& FeatureIDAttribute,
-      int64 VertexIndex);
+      int64 Index);
 };